Improve the edit form feature
Reusing the password protect and edit form forms, although right now
"update" still creates a new paste, although it preserves the attributes
the original (parent) paste had, meaning: the syntax/mimetype and
password (if any).
Where do I wanna go with this? Hopefully showing diffs at some point.
For that I have to take a gander at how to add fields to the Paste
table, migrations are always a pain in the butt in ORMs.