Template directories
This way we point the template directory to the shipped templates (part
of the *package data*) but we can also override the template lookup with
an entry in the ini file pointing to a given location. Dropping files in
this custom location will use these entries instead of the ones shipped
with the package.