Adds an "as_html" option
This option lets the user upload a raw html file for display AS an html file.

This is very strait-forward-- I added a new input to the form and to the
pasttle function, updated the server to handle these appropriately, and updated
the sqlalchemy model with a new boolean column which indicates if it should be
saved as html or not, and an updated to the lexer column so that it can be null
if as_html is true.

Motivation: I would like to have this
https://gist.github.com/d9009b442f5616920d4e someplace where it can be rendered
nicely, but i would have also like to make some references to documents behind
the vpn. If this pr is merged, I can do:

    pandoc ~/dev/booya/nupw/alternative.md -H ~/style.html --standalone | pasttle -r

...and then my work becomes available, nicely styled and formatted, at a pasttle
url.

I ran the tests:

    Ran 13 tests in 0.145s

    OK